+++ src/calcs/Driver.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package postfix;
+package calcs;
 
 /**
  * @author grahamw0
@@ -13,8 +13,16 @@ public class Driver {
    * @param args
    */
   public static void main(String[] args) {
-    PostCalc test = new PostCalc("3 2 + -");
-    //InfixCalc test = new InfixCalc("(((2 + 3))");
+    //PostCalc test = new PostCalc("1 2 3 * 5 + +");
+    InfixCalc test = new InfixCalc("2 * 3 * 1 + 2");
+    
+    
+    /*String testString = "(31 + 234)";
+    String[] testArray = testString.split("(((?<=\\()|(?=\\())|((?<=\\))|(?=\\)))|\\s+)");
+    for(String s : testArray) {
+      System.out.println(s);
+    }*/
+    
     
   }
