+++ src/calcs/PostCalc.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package postfix;
+package calcs;
 
 /**
  * @author grahamw0
